+++ batchglm/train/tf/base_glm/__init__.py
@@ -1,4 +1,5 @@
 from .estimator_graph import GradientGraphGLM, NewtonGraphGLM, TrainerGraphGLM, EstimatorGraphGLM, FullDataModelGraphGLM
-from .jacobians import JacobiansGLM
 from .hessians import HessiansGLM
+from .irls import IRLSGLM
+from .jacobians import JacobiansGLM
 from .model import ESTIMATOR_PARAMS, ProcessModelGLM, ModelVarsGLM, BasicModelGraphGLM
