+++ batchglm/train/tf/glm_nb/external.py
@@ -9,7 +9,7 @@ from batchglm.train.tf.base import TFEstimatorGraph, MonitoredTFEstimator
 
 from batchglm.train.tf.base_glm import GradientGraphGLM, NewtonGraphGLM, TrainerGraphGLM, EstimatorGraphGLM, FullDataModelGraphGLM, BasicModelGraphGLM
 from batchglm.train.tf.base_glm import ESTIMATOR_PARAMS, ProcessModelGLM, ModelVarsGLM
-from batchglm.train.tf.base_glm import HessiansGLM, JacobiansGLM
+from batchglm.train.tf.base_glm import HessiansGLM, IRLSGLM, JacobiansGLM
 
 from batchglm.train.tf.base_glm_all import EstimatorAll, EstimatorGraphAll, HessianTF, JacobiansTF
